requirements.txt
# Core dependencies
pyyaml==6.0.1
feedparser==6.0.10
requests==2.31.0
groq==0.4.1
tweepy==4.14.0

# Additional utilities
python-dateutil==2.8.2
